+++ src/de/earley/pixelengine/entity/Mob.java
@@ -1,8 +1,5 @@
 package de.earley.pixelengine.entity;
 
-import de.earley.pixelengine.level.Level;
-import de.earley.pixelengine.vector.Vector2f;
-
 /**
  *
  * @author timmy
@@ -11,6 +8,6 @@ public abstract class Mob extends Entity {
 
     protected int health;
     protected int damage;
-    protected int speed;
+    protected float speed;
 
 }
